requirements.txt
fastapi
uvicorn
requests

pandas
scikit-learn
joblib

spacy
spacy-transformers
beautifulsoup4

transformers
datasets
evaluate

sentence-transformers
umap-learn
hdbscan
networkx

nltk
tqdm
matplotlib
torch